requirements.txt
google-cloud-sql-connector
pg8000
sqlalchemy
google-cloud-aiplatform
PyPDF2
google-generativeai
python-dotenv